refactor: move route handler to own module
