+++ src/lib/handler/handler.go
@@ -1,0 +1,47 @@
+package handler
+
+import (
+	"appengine"
+	"html/template"
+	"net/http"
+
+	"lib/categories"
+	"lib/constants"
+	"lib/location"
+)
+
+var templates = template.Must(template.ParseGlob("templates/*"))
+
+// RootHandler handles the main call.
+func RootHandler(w http.ResponseWriter, r *http.Request) {
+	c := appengine.NewContext(r)
+
+	selectedLocation := location.Get(r)
+	allCategories := categories.Load(c)
+
+	templateExecuteError := templates.ExecuteTemplate(w, "indexPage", map[string]interface{}{
+		"title":      constants.SiteName,
+		"siteName":   constants.SiteName,
+		"categories": allCategories,
+		"location":   selectedLocation,
+	})
+	if templateExecuteError != nil {
+		c.Errorf("main.RootHandler template: %v", templateExecuteError)
+		return
+	}
+}
+
+// NotFoundHandler handles 404
+func NotFoundHandler(w http.ResponseWriter, r *http.Request) {
+	c := appengine.NewContext(r)
+
+	w.WriteHeader(404)
+	templateExecuteError := templates.ExecuteTemplate(w, "404Page", map[string]interface{}{
+		"title":    "Error 404 – Not found",
+		"siteName": constants.SiteName,
+	})
+	if templateExecuteError != nil {
+		c.Errorf("main.NotFoundHandler template: %v", templateExecuteError)
+		return
+	}
+}
